feat(rust): implement EIP-4788/EIP-2935 system calls and reorganize test files
Add pre-block system calls in L2 block builder using revm's SystemCallEvm
trait so output roots match op-program and kona-host exactly. Split tests
into focused files: simple.rs (derivation correctness), integration.rs
(op-program/kona-host), and server.rs (RPC/beacon API).
