+++ rust/derivation-tests/tests/simple.rs
@@ -7,19 +7,19 @@ use derivation_tests::{config::DeterministicConfig, harness::DerivationTest};
 
 /// Expected super root for 3 empty L2 blocks (deposit-only) derived from empty L1 blocks.
 const EXPECTED_EMPTY_BLOCKS_ROOT: alloy_primitives::B256 =
-    b256!("aacf2b3a7e6b551fee9073326b10fad187b245079da11acdb1a4b77c5c7461db");
+    b256!("19cf64b57f90c792d79c6679d8e67c421f255f07718e2382412b1438315c803e");
 
 /// Expected super root for 1 empty L2 block submitted as a singular batch.
 const EXPECTED_SINGLE_BATCH_ROOT: alloy_primitives::B256 =
-    b256!("f5d6c4e0d16e99712e0304012a3fef20599f3b0ac9e75fa705f58c4c5e53dedd");
+    b256!("3b32a18d695ed075a535131d83a68aed4ae55529ab0241607dd977ad19fe5057");
 
 /// Expected super root for 1 L2 block containing a 1 ETH transfer.
 const EXPECTED_SINGLE_TRANSFER_ROOT: alloy_primitives::B256 =
-    b256!("47c3a6fca4d92dda11dc084f4f8201d51adffe0307536f23bbf4e1f92388c555");
+    b256!("c5281901aaf220b6c3233a66fb08912c8c63543a7f366cf4937993491962aed8");
 
 /// Expected super root for 6 empty L2 blocks (one full epoch).
 const EXPECTED_MULTI_BLOCK_ROOT: alloy_primitives::B256 =
-    b256!("42d497513a28cfa89371cc2c13d44fbbb6934899d18cf56a306b881112318cd0");
+    b256!("beb12ff82b51eda6d1f2621d0e0e2ee8bd35698308a37ad2adcb003bf31d3677");
 
 /// Build a test with N empty L2 blocks derived from N empty L1 blocks.
 fn build_empty_blocks_test(l2_count: usize) -> DerivationTest {
@@ -387,431 +387,4 @@ fn test_system_config_update() {
     assert_eq!(log.data.topics()[1], CONFIG_UPDATE_EVENT_VERSION_0, "topic[1] should be version 0");
 }
 
-/// Integration test: run op-program against empty (deposit-only) L2 blocks.
-///
-/// Verifies that op-program can derive blocks from our test framework's L1/L2 chain.
-/// The claim validation will fail because the Rust framework doesn't replicate
-/// EIP-4788/EIP-2935 system contract state changes — but the derivation pipeline
-/// itself succeeds (block is built, inserted, and safe head advances).
-///
-/// Requires: `OP_PROGRAM_PATH=/path/to/op-program`
-/// Run with: `OP_PROGRAM_PATH=/path/to/op-program cargo test test_op_program_empty_blocks --
-/// --ignored`
-#[tokio::test]
-#[ignore]
-async fn test_op_program_empty_blocks() {
-    use derivation_tests::harness::{run_config_from_test, run_op_program};
-
-    if std::env::var("OP_PROGRAM_PATH").is_err() {
-        eprintln!("SKIP: OP_PROGRAM_PATH not set. Set it to run this integration test.");
-        return;
-    }
-
-    let mut test = DerivationTest::new();
-
-    // Build L1 blocks
-    test.l1.emit_empty_block(); // block 1
-    test.l1.emit_empty_block(); // block 2
-
-    // Build a deposit-only L2 block using L1 genesis as epoch.
-    // The L2 timestamp (genesis + l2_block_time = 1700000002) must be >= the epoch's
-    // timestamp (genesis = 1700000000). L1 block 1 has timestamp 1700000012, which is
-    // too high for the first L2 block.
-    let l1_genesis = test.l1.block_at(0).unwrap().clone();
-    test.l2.set_epoch(&l1_genesis);
-    let block_ref = test.l2.build_empty_block().unwrap();
-
-    // Submit batch referencing L1 genesis as the epoch origin
-    let batch = test.singular_batch_calldata(&[block_ref], &l1_genesis);
-    test.l1.emit_block_with_batches(vec![batch]);
-
-    // The derivation pipeline needs at least one more L1 block after the batch
-    // to advance past it and process the channel.
-    test.l1.emit_empty_block();
-
-    // Start servers — they must stay alive while op-program runs
-    let servers = test.serve().await.unwrap();
-    let config = run_config_from_test(&test, &servers);
-    let rollup_config = test.config.rollup_config();
-    let l1_chain_config = test.config.l1_chain_config();
-
-    let status = run_op_program(&config, &rollup_config, &l1_chain_config)
-        .await
-        .expect("op-program should execute");
-    servers.stop();
-
-    // Op-program exits 1 for claim mismatch (expected — the Rust framework doesn't
-    // replicate EIP-4788/EIP-2935 system contract state changes during block building,
-    // so the output root differs). Derivation success is verified by op-program reaching
-    // the claim validation stage without panicking or hanging.
-    assert!(
-        status.code() == Some(1) || status.success(),
-        "op-program should complete derivation (exit 0 or 1), got: {status}"
-    );
-}
-
-/// Integration test: run kona-host against empty (deposit-only) L2 blocks.
-///
-/// Requires: `KONA_HOST_PATH=/path/to/kona-host`
-/// Run with: `KONA_HOST_PATH=/path/to/kona-host cargo test test_kona_host_empty_blocks --
-/// --ignored`
-#[tokio::test]
-#[ignore]
-async fn test_kona_host_empty_blocks() {
-    use derivation_tests::harness::{run_config_from_test, run_kona_host};
-
-    if std::env::var("KONA_HOST_PATH").is_err() {
-        eprintln!("SKIP: KONA_HOST_PATH not set. Set it to run this integration test.");
-        return;
-    }
-
-    let mut test = DerivationTest::new();
-
-    test.l1.emit_empty_block();
-    test.l1.emit_empty_block();
-
-    // Use L1 genesis as epoch — L2 timestamp must be >= epoch timestamp
-    let l1_genesis = test.l1.block_at(0).unwrap().clone();
-    test.l2.set_epoch(&l1_genesis);
-    let block_ref = test.l2.build_empty_block().unwrap();
-
-    let batch = test.singular_batch_calldata(&[block_ref], &l1_genesis);
-    test.l1.emit_block_with_batches(vec![batch]);
-    test.l1.emit_empty_block();
-
-    let servers = test.serve().await.unwrap();
-    let config = run_config_from_test(&test, &servers);
-    let rollup_config = test.config.rollup_config();
-    let l1_chain_config = test.config.l1_chain_config();
-
-    let status = run_kona_host(&config, &rollup_config, &l1_chain_config)
-        .await
-        .expect("kona-host should execute");
-    servers.stop();
-
-    // Derivation succeeds; claim may mismatch (see op-program tests for explanation).
-    assert!(
-        status.code() == Some(1) || status.success(),
-        "kona-host should complete derivation (exit 0 or 1), got: {status}"
-    );
-}
-
-/// Integration test: run op-program with a user transfer submitted as a singular batch.
-///
-/// Requires: `OP_PROGRAM_PATH=/path/to/op-program`
-/// Run with: `OP_PROGRAM_PATH=/path/to/op-program cargo test test_op_program_with_batch --
-/// --ignored`
-#[tokio::test]
-#[ignore]
-async fn test_op_program_with_batch() {
-    use alloy_consensus::{SignableTransaction, TxEip1559};
-    use alloy_primitives::{Address, TxKind, U256};
-    use alloy_signer::SignerSync;
-    use derivation_tests::harness::{run_config_from_test, run_op_program};
-    use op_alloy_consensus::OpTxEnvelope;
-
-    if std::env::var("OP_PROGRAM_PATH").is_err() {
-        eprintln!("SKIP: OP_PROGRAM_PATH not set. Set it to run this integration test.");
-        return;
-    }
-
-    let mut test = DerivationTest::new();
-
-    test.l1.emit_empty_block();
-    test.l1.emit_empty_block();
-
-    // Use L1 genesis as epoch — L2 timestamp must be >= epoch timestamp
-    let l1_genesis = test.l1.block_at(0).unwrap().clone();
-    test.l2.set_epoch(&l1_genesis);
-
-    // Sign a transfer from the prefunded account
-    let signer = helpers::funded_signer();
-    let tx = TxEip1559 {
-        chain_id: test.config.l2_chain_id,
-        nonce: 0,
-        gas_limit: 21_000,
-        max_fee_per_gas: 0,
-        max_priority_fee_per_gas: 0,
-        to: TxKind::Call(Address::with_last_byte(0x99)),
-        value: U256::from(1_000_000_000_000_000_000u64), // 1 ETH
-        ..Default::default()
-    };
-    let sig = signer.sign_hash_sync(&tx.signature_hash()).expect("signing works");
-    let signed = tx.into_signed(sig);
-    let eth_envelope = alloy_consensus::TxEnvelope::Eip1559(signed);
-    let op_tx = OpTxEnvelope::try_from_eth_envelope(eth_envelope)
-        .expect("should convert ETH envelope to OP envelope");
-
-    let block_ref = test.l2.build_block(vec![op_tx]).unwrap();
-
-    // Submit the batch to L1
-    let batch = test.singular_batch_calldata(&[block_ref], &l1_genesis);
-    test.l1.emit_block_with_batches(vec![batch]);
-    test.l1.emit_empty_block(); // pipeline needs one more L1 block
-
-    let servers = test.serve().await.unwrap();
-    let config = run_config_from_test(&test, &servers);
-    let rollup_config = test.config.rollup_config();
-    let l1_chain_config = test.config.l1_chain_config();
-
-    let status = run_op_program(&config, &rollup_config, &l1_chain_config)
-        .await
-        .expect("op-program should execute");
-    servers.stop();
-
-    // Derivation succeeds; claim may mismatch because the Rust framework doesn't
-    // replicate EIP-4788/EIP-2935 system contract state changes (exit 1 = claim mismatch).
-    assert!(
-        status.code() == Some(1) || status.success(),
-        "op-program should complete derivation (exit 0 or 1), got: {status}"
-    );
-}
-
-/// Integration test: run kona-host with a user transfer submitted as a singular batch.
-///
-/// Requires: `KONA_HOST_PATH=/path/to/kona-host`
-/// Run with: `KONA_HOST_PATH=/path/to/kona-host cargo test test_kona_host_with_batch -- --ignored`
-#[tokio::test]
-#[ignore]
-async fn test_kona_host_with_batch() {
-    use alloy_consensus::{SignableTransaction, TxEip1559};
-    use alloy_primitives::{Address, TxKind, U256};
-    use alloy_signer::SignerSync;
-    use derivation_tests::harness::{run_config_from_test, run_kona_host};
-    use op_alloy_consensus::OpTxEnvelope;
-
-    if std::env::var("KONA_HOST_PATH").is_err() {
-        eprintln!("SKIP: KONA_HOST_PATH not set. Set it to run this integration test.");
-        return;
-    }
-
-    let mut test = DerivationTest::new();
-
-    test.l1.emit_empty_block();
-    test.l1.emit_empty_block();
-
-    // Use L1 genesis as epoch — L2 timestamp must be >= epoch timestamp
-    let l1_genesis = test.l1.block_at(0).unwrap().clone();
-    test.l2.set_epoch(&l1_genesis);
-
-    let signer = helpers::funded_signer();
-    let tx = TxEip1559 {
-        chain_id: test.config.l2_chain_id,
-        nonce: 0,
-        gas_limit: 21_000,
-        max_fee_per_gas: 0,
-        max_priority_fee_per_gas: 0,
-        to: TxKind::Call(Address::with_last_byte(0x99)),
-        value: U256::from(1_000_000_000_000_000_000u64),
-        ..Default::default()
-    };
-    let sig = signer.sign_hash_sync(&tx.signature_hash()).expect("signing works");
-    let signed = tx.into_signed(sig);
-    let eth_envelope = alloy_consensus::TxEnvelope::Eip1559(signed);
-    let op_tx = OpTxEnvelope::try_from_eth_envelope(eth_envelope)
-        .expect("should convert ETH envelope to OP envelope");
-
-    let block_ref = test.l2.build_block(vec![op_tx]).unwrap();
-
-    let batch = test.singular_batch_calldata(&[block_ref], &l1_genesis);
-    test.l1.emit_block_with_batches(vec![batch]);
-    test.l1.emit_empty_block(); // pipeline needs one more L1 block
-
-    let servers = test.serve().await.unwrap();
-    let config = run_config_from_test(&test, &servers);
-    let rollup_config = test.config.rollup_config();
-    let l1_chain_config = test.config.l1_chain_config();
-
-    let status = run_kona_host(&config, &rollup_config, &l1_chain_config)
-        .await
-        .expect("kona-host should execute");
-    servers.stop();
-
-    // Derivation succeeds; claim may mismatch (see op-program tests for explanation).
-    assert!(
-        status.code() == Some(1) || status.success(),
-        "kona-host should complete derivation (exit 0 or 1), got: {status}"
-    );
-}
-
-// ----- Server integration tests -----
-
 mod helpers;
-
-#[tokio::test]
-async fn test_l1_rpc_get_block_by_hash() {
-    let test = build_empty_blocks_test(2);
-    let servers = test.serve().await.unwrap();
-
-    let client = reqwest::Client::new();
-
-    // Get genesis block by number first
-    let resp: serde_json::Value = client
-        .post(servers.l1_rpc_url())
-        .json(&serde_json::json!({
-            "jsonrpc": "2.0",
-            "method": "eth_getBlockByNumber",
-            "params": ["0x0", false],
-            "id": 1
-        }))
-        .send()
-        .await
-        .unwrap()
-        .json()
-        .await
-        .unwrap();
-
-    let block_by_number = resp["result"].clone();
-    let hash = block_by_number["hash"].as_str().unwrap().to_string();
-
-    // Now query by hash
-    let resp: serde_json::Value = client
-        .post(servers.l1_rpc_url())
-        .json(&serde_json::json!({
-            "jsonrpc": "2.0",
-            "method": "eth_getBlockByHash",
-            "params": [hash, false],
-            "id": 2
-        }))
-        .send()
-        .await
-        .unwrap()
-        .json()
-        .await
-        .unwrap();
-
-    let block_by_hash = resp["result"].clone();
-    assert_eq!(
-        block_by_number["hash"], block_by_hash["hash"],
-        "block by number and block by hash should match"
-    );
-    assert_eq!(block_by_number["number"], block_by_hash["number"]);
-
-    servers.stop();
-}
-
-#[tokio::test]
-async fn test_l2_rpc_get_proof() {
-    use derivation_tests::config::L2_TO_L1_MESSAGE_PASSER;
-
-    let test = build_empty_blocks_test(1);
-    let servers = test.serve().await.unwrap();
-
-    let client = reqwest::Client::new();
-
-    let resp: serde_json::Value = client
-        .post(servers.l2_rpc_url())
-        .json(&serde_json::json!({
-            "jsonrpc": "2.0",
-            "method": "eth_getProof",
-            "params": [format!("{:?}", L2_TO_L1_MESSAGE_PASSER), [], "latest"],
-            "id": 1
-        }))
-        .send()
-        .await
-        .unwrap()
-        .json()
-        .await
-        .unwrap();
-
-    let result = &resp["result"];
-    assert!(result.is_object(), "proof response should be an object");
-    assert!(result["address"].is_string(), "proof should have address field");
-    assert!(result["accountProof"].is_array(), "proof should have accountProof");
-    assert!(result["storageProof"].is_array(), "proof should have storageProof");
-
-    servers.stop();
-}
-
-#[tokio::test]
-async fn test_debug_db_get() {
-    let test = build_empty_blocks_test(1);
-    let servers = test.serve().await.unwrap();
-
-    let client = reqwest::Client::new();
-
-    // The state root of the latest L2 block is a known trie node key.
-    let state_root = test.l2.head().header.inner().state_root;
-
-    let resp: serde_json::Value = client
-        .post(servers.l2_rpc_url())
-        .json(&serde_json::json!({
-            "jsonrpc": "2.0",
-            "method": "debug_dbGet",
-            "params": [format!("{state_root:?}")],
-            "id": 1
-        }))
-        .send()
-        .await
-        .unwrap()
-        .json()
-        .await
-        .unwrap();
-
-    // The state root should be a known trie node, so the result should be non-null
-    assert!(
-        resp["result"].is_string(),
-        "debug_dbGet should return data for the state root trie node, got: {resp}"
-    );
-    let result_hex = resp["result"].as_str().unwrap();
-    assert!(result_hex.len() > 2, "result should be non-empty hex data");
-
-    servers.stop();
-}
-
-#[tokio::test]
-async fn test_beacon_blobs() {
-    use derivation_tests::config::DeterministicConfig;
-
-    let config = DeterministicConfig::default();
-    let mut test = derivation_tests::harness::DerivationTest::new();
-
-    // Build L1 blocks
-    test.l1.emit_empty_block(); // block 1
-    test.l1.emit_empty_block(); // block 2
-
-    // Build L2 blocks
-    let l1_block = test.l1.block_at(1).unwrap().clone();
-    test.l2.set_epoch(&l1_block);
-    let block_ref = test.l2.build_empty_block().unwrap();
-
-    // Encode as singular batch calldata
-    let l1_origin = test.l1.block_at(1).unwrap().clone();
-    let batch = test.singular_batch_calldata(&[block_ref], &l1_origin);
-    test.l1.emit_block_with_batches(vec![batch]);
-
-    let servers = test.serve().await.unwrap();
-
-    let client = reqwest::Client::new();
-
-    // Query beacon blobs at slot 0 (genesis) — should return empty data
-    let resp: serde_json::Value = client
-        .get(format!("{}/eth/v1/beacon/blobs/0", servers.beacon_url()))
-        .send()
-        .await
-        .unwrap()
-        .json()
-        .await
-        .unwrap();
-
-    // Verify the response structure
-    assert!(resp["data"].is_array(), "beacon blobs response should have data array");
-    // Without blob batches, data should be empty
-    assert_eq!(resp["data"].as_array().unwrap().len(), 0);
-
-    // Also verify the beacon genesis endpoint works
-    let resp: serde_json::Value = client
-        .get(format!("{}/eth/v1/beacon/genesis", servers.beacon_url()))
-        .send()
-        .await
-        .unwrap()
-        .json()
-        .await
-        .unwrap();
-
-    assert!(resp["data"]["genesis_time"].is_string());
-    let genesis_time: u64 = resp["data"]["genesis_time"].as_str().unwrap().parse().unwrap();
-    assert_eq!(genesis_time, config.genesis_timestamp);
-
-    servers.stop();
-}
